requirements.txt
# Core dependencies for MCP server
fastmcp>=2.12.0
uvicorn>=0.30.0
fastapi>=0.110.0

# MCP protocol
mcp>=1.0.0

# Additional dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Validation libraries
openapi-spec-validator>=0.7.0
jsonschema>=4.0.0

